pyproject.toml
[tool.poetry]
name = "saudi-evisa-data-parser"
version = "0.1.0"
description = "Flask web app to parse Saudi eVisa PDFs into Excel"
authors = ["Your Name <you@example.com>"]
license = "MIT"
package-mode = false  # ✅ ADD THIS LINE


[tool.poetry.dependencies]
python = ">=3.10,<=3.12.1"
flask = "^2.3.3"
pdfplumber = "^0.10.3"
arabic-reshaper = "^3.0.0"
python-bidi = "^0.4.2"
xlsxwriter = "^3.2.0"
gunicorn = "^23.0.0"

[tool.poetry.dev-dependencies]
# optional dev tools
pytest = "^7.4.0"
flake8 = "^6.1.0"

[tool.poetry.scripts]
start-app = "app:app"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"
